refactor: Remove Cookie Sync Interfaces File
