+++ src/sdkRuntimeModels.ts
@@ -36,7 +36,7 @@ import {
 } from './types';
 import IntegrationCapture from './integrationCapture';
 import { INativeSdkHelpers } from './nativeSdkHelpers.interfaces';
-import { ICookieSyncManager, IPixelConfiguration } from './cookieSyncManager.interfaces';
+import { ICookieSyncManager, IPixelConfiguration } from './cookieSyncManager';
 
 // TODO: Resolve this with version in @mparticle/web-sdk
 export type SDKEventCustomFlags = Dictionary<any>;
